refactor(installer): separate platform responsibilities
